x_fase2/week01_day04/am/deployment/backend/requirements.txt
pandas
numpy
joblib
scikit-learn==1.1.3
Flask
tensorflow-cpu
gunicorn